feat(package): drop chat and agents v1 and rename agents_v2 and chat_v2
