+++ archive/v1_code/agents/__init__.py
@@ -1,0 +1,58 @@
+from .base import (
+    Agent,
+    AssistantMessage,
+    ContentDeltaEvent,
+    DoneEvent,
+    ErrorEvent,
+    Field,
+    FunctionCall,
+    Message,
+    StreamEvent,
+    SystemMessage,
+    Tool,
+    ToolCall,
+    ToolCallEvent,
+    ToolMessage,
+    ToolOutput,
+    ToolResponseEvent,
+    UserMessage,
+)
+from .chat_summarization import ChatSummarizationAgent
+from .document_extraction import DocumentExtractionAgent
+from .litigant_assistant import LitigantAssistantAgent
+from .weather import WeatherAgent
+
+# Register agent classes here
+agent_classes = [
+    DocumentExtractionAgent,
+    LitigantAssistantAgent,
+    ChatSummarizationAgent,
+    WeatherAgent,
+]
+agent_registry = {cls.__name__: cls for cls in agent_classes}
+
+
+__all__ = [
+    # Message types
+    "FunctionCall",
+    "ToolCall",
+    "SystemMessage",
+    "UserMessage",
+    "AssistantMessage",
+    "ToolMessage",
+    "Message",
+    # Stream event types
+    "ContentDeltaEvent",
+    "ToolCallEvent",
+    "ToolResponseEvent",
+    "DoneEvent",
+    "ErrorEvent",
+    "StreamEvent",
+    # Core classes
+    "Agent",
+    "Tool",
+    "ToolOutput",
+    "Field",
+    # Registry
+    "agent_registry",
+]
