+++ litigant_portal/agents/__init__.py
@@ -1,58 +1,14 @@
-from .base import (
-    Agent,
-    AssistantMessage,
-    ContentDeltaEvent,
-    DoneEvent,
-    ErrorEvent,
-    Field,
-    FunctionCall,
-    Message,
-    StreamEvent,
-    SystemMessage,
-    Tool,
-    ToolCall,
-    ToolCallEvent,
-    ToolMessage,
-    ToolOutput,
-    ToolResponseEvent,
-    UserMessage,
-)
-from .chat_summarization import ChatSummarizationAgent
-from .document_extraction import DocumentExtractionAgent
-from .litigant_assistant import LitigantAssistantAgent
-from .weather import WeatherAgent
-
-# Register agent classes here
-agent_classes = [
-    DocumentExtractionAgent,
-    LitigantAssistantAgent,
-    ChatSummarizationAgent,
-    WeatherAgent,
-]
-agent_registry = {cls.__name__: cls for cls in agent_classes}
-
+from .assistant import LitigantAssistant
+from .base import Agent, AgentState, Field, Tool, ToolOutput
+from .weather import WeatherAgent, WeatherState
 
 __all__ = [
-    # Message types
-    "FunctionCall",
-    "ToolCall",
-    "SystemMessage",
-    "UserMessage",
-    "AssistantMessage",
-    "ToolMessage",
-    "Message",
-    # Stream event types
-    "ContentDeltaEvent",
-    "ToolCallEvent",
-    "ToolResponseEvent",
-    "DoneEvent",
-    "ErrorEvent",
-    "StreamEvent",
-    # Core classes
     "Agent",
+    "AgentState",
+    "Field",
     "Tool",
     "ToolOutput",
-    "Field",
-    # Registry
-    "agent_registry",
+    "WeatherAgent",
+    "WeatherState",
+    "LitigantAssistant",
 ]
