+++ litigant_portal/agents/base.py
@@ -1,216 +1,44 @@
-import json
-import logging
-from collections.abc import Iterator
-from typing import (
-    TYPE_CHECKING,
-    Annotated,
-    Any,
-    ClassVar,
-    Literal,
-    NotRequired,
-    TypedDict,
-)
-from uuid import UUID
+from typing import Any, ClassVar
 
-from django.conf import settings
 from pydantic import BaseModel, ConfigDict
 from pydantic import Field as PydanticField
-
-if TYPE_CHECKING:
-    from litigant_portal.app.models import ChatSession, UserIdentity
-
-logger = logging.getLogger(__name__)
 
 Field = PydanticField
 
 
-def llm_completion(**kwargs):
-    """Wrapper around litellm.completion() that injects default args."""
-    import litellm
+class AgentState(BaseModel):
+    """Base for an agent's per-thread state model."""
 
-    kwargs.setdefault("drop_params", True)
-    return litellm.completion(**kwargs)
-
-
-# =============================================================================
-# Message Types
-# =============================================================================
-
-
-class FunctionCall(TypedDict, total=False):
-    """Function call details within a tool call."""
-
-    name: str
-    arguments: str
-
-
-class ToolCall(TypedDict, total=False):
-    """A tool call made by the assistant."""
-
-    id: str | None
-    type: Literal["function"]
-    function: FunctionCall
-
-
-class SystemMessage(TypedDict):
-    """A system message that sets the assistant's behavior."""
-
-    role: Literal["system"]
-    content: str
-
-
-class UserMessage(TypedDict):
-    """A message from the user."""
-
-    role: Literal["user"]
-    content: str
-    attachments: NotRequired[list[str]]
-
-
-class AssistantMessage(TypedDict, total=False):
-    """A message from the assistant, optionally with tool calls."""
-
-    role: Literal["assistant"]
-    content: str
-    tool_calls: list[ToolCall]
-
-
-class ToolMessage(TypedDict, total=False):
-    """A response from a tool execution.
-
-    The `data` field stores structured data for the frontend. It is stripped
-    out when sending messages to the LLM API (via messages_for_llm).
-    """
-
-    role: Literal["tool"]
-    tool_call_id: str
-    name: str
-    content: str
-    data: dict[str, Any]
-
-
-class MetaMessage(TypedDict, total=False):
-    """An accounting-only record."""
-
-    role: Literal["meta"]
-    kind: str
-
-
-Message = Annotated[
-    SystemMessage | UserMessage | AssistantMessage | ToolMessage | MetaMessage,
-    Field(discriminator="role"),
-]
-
-
-# =============================================================================
-# Stream Event Types
-# =============================================================================
-
-
-class ContentDeltaEvent(TypedDict):
-    """A chunk of content streamed from the assistant."""
-
-    type: Literal["content_delta"]
-    content: str
-
-
-class ToolCallEvent(TypedDict):
-    """Event indicating the assistant is calling a tool."""
-
-    type: Literal["tool_call"]
-    id: str
-    name: str
-    args: dict[str, Any]
-
-
-class ToolResponseEvent(TypedDict, total=False):
-    """Event containing the result of a tool execution.
-
-    The `data` field contains structured data for the frontend.
-    The `response` field is omitted from streaming by default.
-    """
-
-    type: Literal["tool_response"]
-    id: str
-    name: str
-    data: dict[str, Any]
-
-
-class DoneEvent(TypedDict):
-    """Event indicating the stream is complete."""
-
-    type: Literal["done"]
-
-
-class ErrorEvent(TypedDict):
-    """Event indicating an error occurred."""
-
-    type: Literal["error"]
-    error: str
-
-
-StreamEvent = (
-    ContentDeltaEvent
-    | ToolCallEvent
-    | ToolResponseEvent
-    | DoneEvent
-    | ErrorEvent
-)
-
-
-# =============================================================================
-# Agent Classes
-# =============================================================================
+    model_config = ConfigDict(extra="allow")
 
 
 class ToolOutput(BaseModel):
-    """Output from a tool execution.
+    """Result of running a tool."""
 
-    Tools can return either a string (shorthand for just response) or a
-    ToolOutput with both response (for the LLM) and data (for the frontend).
-
-    Example:
-        def __call__(self, agent: Agent) -> ToolOutput:
-            return ToolOutput(
-                response=f"Weather in {self.location}: 72F, sunny",
-                data={"location": self.location, "temp_f": 72, "condition": "sunny"}
-            )
-    """
-
-    response: str
-    data: dict[str, Any] | None = None
+    # Returned to the model as the tool message content.
+    result: str
+    # Arbitrary structured data streamed to the frontend for rendering.
+    render_data: dict[str, Any] | None = None
+    # When True, the agent regenerates its system prompt before the next
+    # step (instead of only when a new user message is submitted).
+    refresh_system_prompt: bool = False
+    # LLM spend the tool incurred itself (e.g. a sub-model call); the
+    # engine books it on the tool response message so thread usage
+    # reflects the true total.
+    cost: float = 0.0
 
 
 class Tool(BaseModel):
-    """Base class for agent tools.
+    """Base class for agent tools."""
 
-    Extend this class and define fields as tool parameters.
+    model_config = ConfigDict(extra="allow")
 
-    Example:
-        class WeatherTool(Tool):
-            '''Get the current weather for a location.'''
-            location: str = Field(description="City name")
+    tool_call_template: ClassVar[str | bool | None] = None
+    tool_result_template: ClassVar[str | bool | None] = None
 
-            def __call__(self, agent: Agent) -> ToolOutput:
-                return ToolOutput(
-                    response=f"Weather in {self.location}: 72F, sunny",
-                    data={"location": self.location, "temp_f": 72}
-                )
-
-    You can also return a plain string for simple cases:
-
-        def __call__(self, agent: Agent) -> str:
-            return f"Weather in {self.location}: 72F, sunny"
-    """
-
-    def __call__(self, agent: "Agent") -> ToolOutput | str | None:
-        """Execute the tool.
-
-        Override this method to implement tool logic.
-        Return a ToolOutput for structured data, or a string for simple responses.
-        You can also store data in agent.state.
-        """
-        pass
+    def __call__(self, *, thread_id) -> ToolOutput:
+        """Run the tool against ``thread_id`` and return its output."""
+        raise NotImplementedError
 
     @classmethod
     def get_schema(cls) -> dict:
@@ -219,361 +47,29 @@ class Tool(BaseModel):
             "type": "function",
             "function": {
                 "name": cls.__name__,
-                "description": cls.__doc__,
+                "description": cls.__doc__ or "",
                 "parameters": cls.model_json_schema(),
             },
         }
 
-    model_config = ConfigDict(extra="allow")
-
 
 class Agent:
-    """A litellm wrapper for tool-calling agents with streaming support.
+    """A configuration object for a tool-calling agent."""
 
-    Subclass this to create domain-specific agents with default configurations.
-    All messages are plain dicts at runtime for easy serialization.
+    completion_args: ClassVar[dict[str, Any]] = {}
+    state_schema: ClassVar[type[AgentState]] = AgentState
+    tools: ClassVar[list[type[Tool]]] = []
 
-    Example:
-        class LegalAssistant(Agent):
-            default_tools = [SearchDocuments, GetCourtInfo]
-            default_messages = [
-                {"role": "system", "content": "You are a helpful legal assistant..."}
-            ]
-
-    The model is read from the CHAT_MODEL env var, defaulting to
-    "openai/gpt-4o-mini".
-    """
-
-    default_model: ClassVar[str] = ""
-    default_tools: ClassVar[list[type[Tool]]] = []
-    default_max_steps: ClassVar[int] = 30
-    default_messages: ClassVar[list[Message]] = []
-    default_completion_args: ClassVar[dict[str, Any]] = {}
-
-    def __init__(
-        self,
-        model: str | None = None,
-        tools: list[type[Tool]] | None = None,
-        messages: list[Message] | None = None,
-        state: dict | None = None,
-        max_steps: int | None = None,
-        session: type["ChatSession"] | None = None,
-        **completion_args: Any,
-    ):
-        """Initialize the agent.
-
-        Args:
-            model: LiteLLM model string. Defaults to CHAT_MODEL env var.
-            tools: List of Tool classes available to the agent.
-            messages: Initial message history (dicts matching Message schema).
-            state: Initial state dictionary for tool data storage.
-            max_steps: Maximum number of agent steps (tool call loops).
-            session: ChatSession object to associate with the agent.
-            **completion_args: Additional args passed to litellm.completion().
-        """
-        self.session = session
-        self.model = model or self.default_model or settings.CHAT_MODEL
-        self.completion_args = {
-            **self.default_completion_args,
-            **completion_args,
-        }
-
-        # Set up tools
-        tool_classes = tools if tools is not None else self.default_tools
-        self.tools = {tool.__name__: tool for tool in tool_classes}
-        self.tool_schemas = [
-            tool.get_schema() for tool in tool_classes
-        ] or None
-
-        # Set up messages
-        if messages is not None:
-            self.messages: list[Message] = list(messages)
-        else:
-            self.messages = list(self.default_messages)
-
-        self.state = state or {}
-        self.max_steps = max_steps or self.default_max_steps
-        self.last_response = None
+    def generate_system_prompt(self, *, thread_id) -> str:
+        """Build the system prompt for ``thread_id`` from its state."""
+        raise NotImplementedError
 
     @property
-    def messages_for_llm(self) -> list[dict[str, Any]]:
-        """Return messages sanitized for LLM API calls.
+    def tools_by_name(self) -> dict[str, type[Tool]]:
+        """Map tool name -> tool class for dispatching tool calls."""
+        return {tool.__name__: tool for tool in self.tools}
 
-        Strips out fields like `data` that aren't valid for the API.
-        """
-        messages: list[dict[str, Any]] = []
-        for msg in self.messages:
-            if msg.get("role") == "tool" and "data" in msg:
-                messages.append({k: v for k, v in msg.items() if k != "data"})
-            else:
-                messages.append(dict(msg))
-        return messages
-
-    def add_message(self, message: Message):
-        """Add a message to the agent's message history."""
-        from litigant_portal.app.models import Message as MessageModel
-
-        self.messages.append(message)
-        if self.session:
-            MessageModel.objects.create(session=self.session, data=message)
-
-    def execute_tool(
-        self, tool_call: ToolCall
-    ) -> tuple[ToolMessage, ToolResponseEvent]:
-        """Execute a single tool call.
-
-        Returns:
-            A tuple of (tool_message, tool_response_event):
-            - tool_message: The full message to store (includes data)
-            - tool_response_event: The event to stream to frontend (data only)
-        """
-        function = tool_call.get("function", {})
-        tool_name = function.get("name", "")
-        tool_id = tool_call.get("id", "") or ""
-        tool_class = self.tools.get(tool_name)
-
-        try:
-            if not tool_class:
-                raise ValueError(f"Unknown tool: {tool_name}")
-
-            tool_args = json.loads(function.get("arguments", "{}"))
-            tool_instance = tool_class(**tool_args)
-            result = tool_instance(self)
-
-            # Normalize result to ToolOutput
-            if result is None:
-                output = ToolOutput(response="Success")
-            elif isinstance(result, str):
-                output = ToolOutput(response=result)
-            elif isinstance(result, ToolOutput):
-                output = result
-            else:
-                output = ToolOutput(**result)
-
-        except Exception as e:
-            logger.error(f"Tool {tool_name} failed: {e}")
-            output = ToolOutput(response=f"Error: {e}")
-
-        # Build the tool message
-        tool_msg: ToolMessage = {
-            "role": "tool",
-            "tool_call_id": tool_id,
-            "name": tool_name,
-            "content": output.response,
-        }
-        if output.data is not None:
-            tool_msg["data"] = output.data
-
-        # Build the event
-        response_event: ToolResponseEvent = {
-            "type": "tool_response",
-            "id": tool_id,
-            "name": tool_name,
-        }
-        if output.data is not None:
-            response_event["data"] = output.data
-
-        return tool_msg, response_event
-
-    def stream_run(
-        self,
-        messages: list[Message] | str | None = None,
-        **completion_args: Any,
-    ) -> Iterator[StreamEvent]:
-        """Run agent loop with streaming, yielding events as they occur.
-
-        Args:
-            messages: Initial user message(s) to add to history.
-            **completion_args: Override completion arguments.
-
-        Yields:
-            Event dicts for frontend consumption.
-        """
-        # Add initial messages
-        if messages is not None:
-            if isinstance(messages, str):
-                user_msg: UserMessage = {"role": "user", "content": messages}
-                self.add_message(user_msg)
-            else:
-                for msg in messages:
-                    self.add_message(msg)
-
-        try:
-            for _ in range(self.max_steps):
-                # Prepare completion args
-                call_args = {
-                    "model": self.model,
-                    "messages": self.messages_for_llm,
-                    "tools": self.tool_schemas,
-                    "stream": True,
-                    **self.completion_args,
-                    **completion_args,
-                }
-
-                # Stream the response
-                full_content: list[str] = []
-                tool_calls: list[ToolCall] = []
-
-                for chunk in llm_completion(**call_args):
-                    choice = chunk.choices[0]
-                    delta = choice.delta
-
-                    # Yield content deltas
-                    if delta.content:
-                        full_content.append(delta.content)
-                        content_event: ContentDeltaEvent = {
-                            "type": "content_delta",
-                            "content": delta.content,
-                        }
-                        yield content_event
-
-                    # Accumulate tool calls from stream
-                    if delta.tool_calls:
-                        for tc in delta.tool_calls:
-                            # Extend tool_calls list if needed
-                            while tc.index >= len(tool_calls):
-                                tool_calls.append(
-                                    {
-                                        "id": None,
-                                        "type": "function",
-                                        "function": {
-                                            "name": "",
-                                            "arguments": "",
-                                        },
-                                    }
-                                )
-                            # Update the tool call at this index
-                            if tc.id:
-                                tool_calls[tc.index]["id"] = tc.id
-                            if tc.function:
-                                if tc.function.name:
-                                    tool_calls[tc.index]["function"][
-                                        "name"
-                                    ] = tc.function.name
-                                if tc.function.arguments:
-                                    tool_calls[tc.index]["function"][
-                                        "arguments"
-                                    ] += tc.function.arguments
-
-                # Store assistant message
-                assistant_msg: AssistantMessage = {
-                    "role": "assistant",
-                    "content": "".join(full_content),
-                }
-                if tool_calls:
-                    assistant_msg["tool_calls"] = tool_calls
-                self.add_message(assistant_msg)
-
-                # If we have tool calls, yield them and execute
-                if tool_calls:
-                    for tool_call in tool_calls:
-                        function = tool_call.get("function", {})
-                        # Yield tool call event
-                        try:
-                            args_dict = json.loads(
-                                function.get("arguments", "{}")
-                            )
-                        except json.JSONDecodeError:
-                            args_dict = {}
-
-                        tool_call_event: ToolCallEvent = {
-                            "type": "tool_call",
-                            "id": tool_call.get("id") or "",
-                            "name": function.get("name", ""),
-                            "args": args_dict,
-                        }
-                        yield tool_call_event
-
-                        # Execute tool
-                        tool_msg, tool_response_event = self.execute_tool(
-                            tool_call
-                        )
-
-                        # Store message (with data) and yield event (data only)
-                        self.add_message(tool_msg)
-                        yield tool_response_event
-                else:
-                    break
-
-            done_event: DoneEvent = {"type": "done"}
-            yield done_event
-
-        except Exception as e:
-            logger.error(f"Agent error: {e}")
-            error_event: ErrorEvent = {"type": "error", "error": str(e)}
-            yield error_event
-
-    def ping(self) -> bool:
-        """Check if the agent's model/provider is accessible."""
-        try:
-            llm_completion(
-                model=self.model,
-                messages=[{"role": "user", "content": "hi"}],
-                max_tokens=1,
-            )
-            return True
-        except Exception as e:
-            logger.warning(f"Health check failed for {self.model}: {e}")
-            return False
-
-    @classmethod
-    def from_session_id(
-        cls,
-        identity: "UserIdentity",
-        session_id: str | UUID | None = None,
-        **kwargs: Any,
-    ) -> "Agent":
-        """Create an agent from a session ID.
-
-        If no session ID is provided, a new session is created.
-        """
-        from litigant_portal.app.models import ChatSession
-        from litigant_portal.app.models import Message as MessageModel
-
-        if not session_id:
-            topic = kwargs.pop("topic", "")
-            jurisdiction = kwargs.pop("jurisdiction", "")
-            court = kwargs.pop("court", None)
-            phase = kwargs.pop("phase", None)
-            session = ChatSession.objects.create(
-                identity=identity,
-                topic=topic,
-                jurisdiction=jurisdiction,
-            )
-            # New sessions start in triage unless explicitly overridden.
-            kwargs.setdefault("phase", phase or "triage")
-            agent = cls(
-                session=session,
-                topic=topic,
-                court=court,
-                jurisdiction=jurisdiction,
-                **kwargs,
-            )
-            for msg in agent.messages:
-                MessageModel.objects.create(session=session, data=msg)
-            return agent
-        else:
-            try:
-                session = ChatSession.objects.get(id=session_id)
-            except ChatSession.DoesNotExist:
-                raise ValueError(f"Session {session_id} not found")
-            if session.identity_id != identity.pk:
-                raise PermissionError("Unauthorized access to session")
-            messages = [
-                m.data for m in session.messages.order_by("created_at")
-            ]
-            # Restore topic/jurisdiction from session for prompt composition
-            if session.topic:
-                kwargs.setdefault("topic", session.topic)
-            if session.jurisdiction:
-                kwargs.setdefault("jurisdiction", session.jurisdiction)
-            # Derive current phase from session state (see litigant_portal.prompts.phase_for_session)
-            from litigant_portal.prompts import phase_for_session
-
-            kwargs.setdefault("phase", phase_for_session(session))
-            agent = cls(session=session, messages=messages, **kwargs)
-            return agent
-
-    def __call__(self, *args: Any, **kwargs: Any) -> Any:
-        """Hook for custom workflow logic."""
-        pass
+    @property
+    def tool_schemas(self) -> list[dict] | None:
+        """Tool schemas for the LLM, or None when the agent has no tools."""
+        return [tool.get_schema() for tool in self.tools] or None
