+++ litigant_portal/app/models/__init__.py
@@ -1,5 +1,4 @@
 from .admin import *  # noqa: F403
 from .assistant import *  # noqa: F403
-from .chat import *  # noqa: F403
-from .chat_v2 import *  # noqa: F403
+from .chat_engine import *  # noqa: F403
 from .user import *  # noqa: F403
