+++ litigant_portal/app/templates/base.html
@@ -39,10 +39,8 @@
     <script src="{% static 'js/theme.js' %}"></script>
     {# Alpine.data() components - must load before Alpine #}
     <script src="{% static 'js/components.js' %}"></script>
-    {# Chat component - must load before Alpine #}
-    <script src="{% static 'js/chat.js' %}"></script>
-    {# Chat v2 helpers #}
-    <script src="{% static 'js/chat_v2.js' %}"></script>
+    {# Chat engine helpers #}
+    <script src="{% static 'js/chat_engine.js' %}"></script>
     {# Alpine.js CSP build - required for Content Security Policy compliance #}
     {# https://alpinejs.dev/advanced/csp #}
     {# Uses non-minified version in debug mode for easier debugging #}
