--- litigant_portal/app/views/__init__.py
@@ -1,2 +1,0 @@
-from .endpoints import *  # noqa: F403
-from .pages import *  # noqa: F403
